Add conditional to run update scripts only if needed

+++ modules/prestretch/templates/delete_onlyoffice_image.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-#delete onlyoffice docker image
-docker rmi -f $(docker images | grep onlyoffice/documentserver | awk '{ print $3 }')
+#if onlyoffice image exists
+if docker images | grep onlyoffice/documentserver; then
+
+  #delete onlyoffice docker image
+  docker rmi -f $(docker images | grep onlyoffice/documentserver | awk '{ print $3 }')
+
+fi
