+++ modules/prestretch/templates/update_postgresql_96.sh
@@ -1,31 +1,39 @@
 #!/bin/bash
 
-#update mongodb to 9.6
-apt-get -t jessie-backports install postgresql-9.6 postgresql-client-9.6 postgresql-server-dev-9.6 -y
+#if postgresql version is 9.4
+if pg_config --version | grep 9.4; then
 
-#update cluster to 9.6 and disable previous cluster
-#with workaround if there are client connections active
-#doc: https://gist.github.com/johanndt/6436bfad28c86b28f794
-sleep 10
-pg_dropcluster 9.6 main --stop
-sleep 5
-pg_upgradecluster 9.4 main
-sleep 5
-service postgresql restart
-sleep 5
-service postgresql stop
-sleep 5
-pg_upgradecluster 9.4 main
-sleep 10
-pg_dropcluster 9.4 main
-sleep 10
+  echo "## Update postgresql to 9.6 ##################################"
 
-#purge old packages
-apt-get --purge remove postgresql-9.4 postgresql-client-9.4 postgresql-server-dev-9.4 -y
+  #update postgresql to 9.6
+  apt-get -t jessie-backports install postgresql-9.6 postgresql-client-9.6 postgresql-server-dev-9.6 -y
 
-#restart postgresql
-service postgresql restart
+  #update cluster to 9.6 and disable previous cluster
+  #with workaround if there are client connections active
+  #doc: https://gist.github.com/johanndt/6436bfad28c86b28f794
+  sleep 10
+  pg_dropcluster 9.6 main --stop
+  sleep 5
+  pg_upgradecluster 9.4 main
+  sleep 5
+  service postgresql restart
+  sleep 5
+  service postgresql stop
+  sleep 5
+  pg_upgradecluster 9.4 main
+  sleep 10
+  pg_dropcluster 9.4 main
+  sleep 10
 
-#list clusters
-pg_lsclusters
+  #purge old packages
+  apt-get --purge remove postgresql-9.4 postgresql-client-9.4 postgresql-server-dev-9.4 -y
+
+  #restart postgresql
+  service postgresql restart
+
+  #list clusters
+  pg_lsclusters
+
+fi
+
