+++ modules/prestretch/templates/upgrade_stretch.sh
@@ -5,10 +5,11 @@ apt-get update
 apt list --upgradable
 apt-get upgrade -y
 
-#delete root /.my.conf
-rm /root/.my.cnf
+if [ -f /root/.my.cnf ]; then
+  #delete root /.my.conf
+  rm /root/.my.cnf
+fi
 
 #dist-upgrade stretch
 apt-get dist-upgrade -y
 
-#TODO, check exit code of dpkg commands
